#47 Removed base component and shifted the class to the plugin module.
