--- softvis3d-base/README.md
@@ -1,3 +1,0 @@
-# SoftVis3D - Base package
-
-Contains base java files to get the visualization data used by the view.
