+++ sonar-softvis3d-plugin/src/main/java/de/rinderle/softvis3d/domain/tree/TreeNodeType.java
@@ -17,31 +17,8 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package de.rinderle.softvis3d.base.domain;
+package de.rinderle.softvis3d.domain.tree;
 
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class MetricTest {
-
-  @Test
-  public void testGetter() throws Exception {
-    final Integer id = 12213;
-    final String description = "testDescription";
-    final Metric metric = new Metric(id, description);
-
-    assertEquals(id, metric.getId());
-    assertEquals(description, metric.getDescription());
-  }
-
-  @Test(expected = IllegalArgumentException.class)
-  public void testNullId() throws Exception {
-    new Metric(null, "testDescription");
-  }
-
-  @Test(expected = IllegalArgumentException.class)
-  public void testNullDescription() throws Exception {
-    new Metric(123, null);
-  }
+public enum TreeNodeType {
+  TREE, PATH_GENERATED
 }
