+++ sonar-softvis3d-plugin/src/main/java/de/rinderle/softvis3d/preprocessing/PreProcessor.java
@@ -21,10 +21,10 @@ package de.rinderle.softvis3d.preprocessing;
 
 import com.google.inject.Inject;
 import de.rinderle.softvis3d.SoftVis3DPlugin;
-import de.rinderle.softvis3d.base.domain.tree.RootTreeNode;
 import de.rinderle.softvis3d.cache.SnapshotCacheService;
 import de.rinderle.softvis3d.domain.SnapshotStorageKey;
 import de.rinderle.softvis3d.domain.VisualizationRequest;
+import de.rinderle.softvis3d.domain.tree.RootTreeNode;
 import de.rinderle.softvis3d.preprocessing.tree.OptimizeTreeStructure;
 import de.rinderle.softvis3d.preprocessing.tree.TreeBuilder;
 import org.sonar.api.server.ws.LocalConnector;
