+++ sonar-softvis3d-plugin/src/main/java/de/rinderle/softvis3d/preprocessing/tree/TreeBuilder.java
@@ -20,10 +20,10 @@
 package de.rinderle.softvis3d.preprocessing.tree;
 
 import com.google.inject.Inject;
-import de.rinderle.softvis3d.base.domain.tree.RootTreeNode;
 import de.rinderle.softvis3d.dao.DaoService;
 import de.rinderle.softvis3d.domain.VisualizationRequest;
 import de.rinderle.softvis3d.domain.sonar.SonarMeasure;
+import de.rinderle.softvis3d.domain.tree.RootTreeNode;
 
 import java.util.List;
