+++ sonar-softvis3d-plugin/src/main/java/de/rinderle/softvis3d/webservice/SoftVis3dJsonWriter.java
@@ -17,9 +17,10 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package de.rinderle.softvis3d.base.result;
+package de.rinderle.softvis3d.webservice;
 
 import com.google.gson.stream.JsonWriter;
+
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
