+++ sonar-softvis3d-plugin/src/main/java/de/rinderle/softvis3d/webservice/VisualizationWebserviceHandler.java
@@ -20,11 +20,9 @@
 package de.rinderle.softvis3d.webservice;
 
 import com.google.inject.Inject;
-import de.rinderle.softvis3d.base.domain.tree.RootTreeNode;
-import de.rinderle.softvis3d.base.result.SoftVis3dJsonWriter;
-import de.rinderle.softvis3d.base.result.TreeNodeJsonWriter;
 import de.rinderle.softvis3d.dao.DaoService;
 import de.rinderle.softvis3d.domain.VisualizationRequest;
+import de.rinderle.softvis3d.domain.tree.RootTreeNode;
 import de.rinderle.softvis3d.preprocessing.PreProcessor;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
