+++ sonar-softvis3d-plugin/src/test/java/de/rinderle/softvis3d/domain/tree/TreeNodeTest.java
@@ -1,4 +1,4 @@
-package de.rinderle.softvis3d.base.domain.tree;
+package de.rinderle.softvis3d.domain.tree;
 
 import org.junit.Test;
