+++ sonar-softvis3d-plugin/src/test/java/de/rinderle/softvis3d/preprocessing/PreProcessorTest.java
@@ -19,11 +19,11 @@
  */
 package de.rinderle.softvis3d.preprocessing;
 
-import de.rinderle.softvis3d.base.domain.tree.RootTreeNode;
 import de.rinderle.softvis3d.cache.SnapshotCacheService;
 import de.rinderle.softvis3d.dao.DaoService;
 import de.rinderle.softvis3d.domain.SnapshotStorageKey;
 import de.rinderle.softvis3d.domain.VisualizationRequest;
+import de.rinderle.softvis3d.domain.tree.RootTreeNode;
 import de.rinderle.softvis3d.preprocessing.tree.OptimizeTreeStructure;
 import de.rinderle.softvis3d.preprocessing.tree.TreeBuilder;
 import org.junit.Before;
