+++ sonar-softvis3d-plugin/src/test/java/de/rinderle/softvis3d/webservice/SoftVis3dJsonWriterTest.java
@@ -17,7 +17,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package de.rinderle.softvis3d.base.result;
+package de.rinderle.softvis3d.webservice;
 
 import java.io.IOException;
 import java.io.OutputStream;
