+++ sonar-softvis3d-plugin/src/test/java/de/rinderle/softvis3d/webservice/TreeNodeJsonWriterTest.java
@@ -17,10 +17,10 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package de.rinderle.softvis3d.base.result;
+package de.rinderle.softvis3d.webservice;
 
-import de.rinderle.softvis3d.base.TestTreeBuilder;
-import de.rinderle.softvis3d.base.domain.tree.RootTreeNode;
+import de.rinderle.softvis3d.TestTreeBuilder;
+import de.rinderle.softvis3d.domain.tree.RootTreeNode;
 
 import java.io.IOException;
 import java.io.OutputStream;
